Adapt to moved headers in opm-grid.
