+++ examples/mimetic_periodic_test.cpp
@@ -38,7 +38,7 @@
 #include <opm/parser/eclipse/Units/Units.hpp>
 #include <opm/common/utility/parameters/ParameterGroup.hpp>
 
-#include <dune/grid/CpGrid.hpp>
+#include <opm/grid/CpGrid.hpp>
 
 #include <opm/porsol/common/PeriodicHelpers.hpp>
 #include <opm/porsol/common/BoundaryConditions.hpp>
