+++ opm/elasticity/elasticity_preconditioners.hpp
@@ -29,7 +29,7 @@
 
 #include <opm/common/utility/platform_dependent/reenable_warnings.h>
 
-#include <dune/grid/CpGrid.hpp>
+#include <opm/grid/CpGrid.hpp>
 #include <opm/elasticity/asmhandler.hpp>
 #include <opm/elasticity/matrixops.hpp>
