+++ opm/porsol/common/Wells.hpp
@@ -21,7 +21,7 @@
 #define OPM_WELLS_HEADER_INCLUDED
 
 #include <opm/common/ErrorMacros.hpp>
-#include <opm/core/utility/SparseTable.hpp>
+#include <opm/grid/utility/SparseTable.hpp>
 #include <opm/parser/eclipse/Deck/Deck.hpp>
 #include <dune/common/fvector.hh>
