+++ opm/porsol/euler/EulerUpstream_impl.hpp
@@ -42,9 +42,9 @@
 #include <opm/common/ErrorMacros.hpp>
 #include <opm/core/utility/Average.hpp>
 #include <opm/parser/eclipse/Units/Units.hpp>
-#include <dune/grid/common/Volumes.hpp>
+#include <opm/grid/common/Volumes.hpp>
 #include <opm/porsol/euler/CflCalculator.hpp>
-#include <opm/core/utility/StopWatch.hpp>
+#include <opm/grid/utility/StopWatch.hpp>
 
 #include <cassert>
 #include <cmath>
