+++ opm/porsol/mimetic/MimeticIPEvaluator.hpp
@@ -43,7 +43,7 @@
 #include <array>
 
 #include <opm/common/ErrorMacros.hpp>
-#include <opm/core/utility/SparseTable.hpp>
+#include <opm/grid/utility/SparseTable.hpp>
 
 #include <opm/porsol/common/fortran.hpp>
 #include <opm/porsol/common/blas_lapack.hpp>
