+++ opm/upscaling/SteadyStateUpscalerManager.hpp
@@ -40,7 +40,7 @@
 #include <opm/upscaling/SteadyStateUpscaler.hpp>
 #include <opm/upscaling/UpscalingTraits.hpp>
 #include <opm/parser/eclipse/Units/Units.hpp>
-#include <opm/core/utility/SparseTable.hpp>
+#include <opm/grid/utility/SparseTable.hpp>
 #include <cmath>
 #include <fstream>
 #include <iostream>
