+++ opm/upscaling/initCPGrid.cpp
@@ -40,7 +40,7 @@
 #include <opm/parser/eclipse/Parser/ParseContext.hpp>
 #include <opm/parser/eclipse/Deck/Deck.hpp>
 
-#include <dune/grid/CpGrid.hpp>
+#include <opm/grid/CpGrid.hpp>
 
 #include <opm/common/utility/parameters/ParameterGroup.hpp>
